+++ src/module/applications/sheets/item-sheet.mjs
@@ -259,7 +259,8 @@ export default class DrawSteelItemSheet extends DSDocumentSheetMixin(sheets.Item
     };
 
     // Iterate over active effects, classifying them into categories
-    for (const e of this.item.effects) {
+    const effects = this.item.effects.contents.sort((a, b) => a.sort - b.sort);
+    for (const e of effects) {
       const effectContext = {
         id: e.id,
         uuid: e.uuid,
@@ -787,7 +788,7 @@ export default class DrawSteelItemSheet extends DSDocumentSheetMixin(sheets.Item
     }
 
     // Perform the sort
-    const sortUpdates = SortingHelpers.performIntegerSort(effect, {
+    const sortUpdates = foundry.utils.performIntegerSort(effect, {
       target,
       siblings,
     });